+++ apps/aptos/components/Liquidity/components/LiquidityBridgeWarning.tsx
@@ -1,26 +1,11 @@
-import { useMemo } from 'react'
-import { useAccount, useAccountBalance } from '@pancakeswap/awgmi'
 import { Currency } from '@pancakeswap/aptos-swap-sdk'
 import useBridgeInfo from 'components/Swap/hooks/useBridgeInfo'
 import { Message, Flex, Text, Link } from '@pancakeswap/uikit'
 import { useTranslation } from '@pancakeswap/localization'
 
 const LiquidityBridgeWarning = ({ currency }: { currency?: Currency }) => {
-  const { account } = useAccount()
   const { t } = useTranslation()
-  const bridgeResult = useBridgeInfo({ currency })
-
-  const { data, isLoading } = useAccountBalance({
-    address: account?.address,
-    coin: currency?.wrapped?.address,
-    enabled: !!currency,
-    watch: true,
-  })
-
-  const showBridgeWarning = useMemo(
-    () => bridgeResult && !isLoading && Number(data?.formatted) <= 0,
-    [data, isLoading, bridgeResult],
-  )
+  const { showBridgeWarning, bridgeResult } = useBridgeInfo({ currency })
 
   return (
     <>
@@ -41,7 +26,7 @@ const LiquidityBridgeWarning = ({ currency }: { currency?: Currency }) => {
               {bridgeResult?.platform}
             </Link>
             <Text fontSize="12px" color="warning" m="auto">
-              {t('to bridge this asset.')}
+              {t('to bridge %symbol%.', { symbol: currency?.symbol ?? '' })}
             </Text>
           </Flex>
         </Message>
